+++ src/UI/FilterBar.js
@@ -1,5 +1,5 @@
-import { $ } from "../../utils/Dom";
-import { sortByName, sortByDistance } from "../../utils/Sort";
+import { $ } from "../utils/Dom";
+import { sortByName, sortByDistance } from "../utils/Sort";
 
 export default class FilterBar {
   #template = `
